:recycle: Refactor: network -> core/network로 폴더 변경
