+++ composeApp/src/commonMain/kotlin/com/dh/ondot/core/di/ServiceLocator.kt
@@ -12,8 +12,8 @@ import com.dh.ondot.domain.service.AlarmScheduler
 import com.dh.ondot.domain.service.AlarmStorage
 import com.dh.ondot.domain.service.MapProviderStorage
 import com.dh.ondot.domain.service.SoundPlayer
-import com.dh.ondot.network.NetworkClient
-import com.dh.ondot.network.TokenProvider
+import com.dh.ondot.core.network.NetworkClient
+import com.dh.ondot.core.network.TokenProvider
 
 object ServiceLocator {
     private lateinit var tokenProvider: TokenProvider
