+++ composeApp/src/commonMain/kotlin/com/dh/ondot/core/network/HttpMethod.kt
@@ -1,4 +1,4 @@
-package com.dh.ondot.network
+package com.dh.ondot.core.network
 
 enum class HttpMethod {
     GET, POST, PUT, PATCH, DELETE
