+++ composeApp/src/commonMain/kotlin/com/dh/ondot/core/network/NetworkConfig.kt
@@ -1,0 +1,3 @@
+package com.dh.ondot.core.network
+
+const val BASE_URL = "https://on-dot.site"
