+++ composeApp/src/commonMain/kotlin/com/dh/ondot/data/repository/PlaceRepositoryImpl.kt
@@ -2,8 +2,8 @@ package com.dh.ondot.data.repository
 
 import com.dh.ondot.domain.model.response.AddressInfo
 import com.dh.ondot.domain.repository.PlaceRepository
-import com.dh.ondot.network.HttpMethod
-import com.dh.ondot.network.NetworkClient
+import com.dh.ondot.core.network.HttpMethod
+import com.dh.ondot.core.network.NetworkClient
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
