+++ composeApp/src/commonMain/kotlin/com/dh/ondot/data/repository/ScheduleRepositoryImpl.kt
@@ -7,8 +7,8 @@ import com.dh.ondot.domain.model.response.ScheduleAlarmResponse
 import com.dh.ondot.domain.model.response.ScheduleDetail
 import com.dh.ondot.domain.model.response.ScheduleListResponse
 import com.dh.ondot.domain.repository.ScheduleRepository
-import com.dh.ondot.network.HttpMethod
-import com.dh.ondot.network.NetworkClient
+import com.dh.ondot.core.network.HttpMethod
+import com.dh.ondot.core.network.NetworkClient
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
