+++ composeApp/src/iosMain/kotlin/com/dh/ondot/core/di/PlatformDependencies.ios.kt
@@ -15,7 +15,7 @@ import com.dh.ondot.domain.service.AlarmScheduler
 import com.dh.ondot.domain.service.AlarmStorage
 import com.dh.ondot.domain.service.MapProviderStorage
 import com.dh.ondot.domain.service.SoundPlayer
-import com.dh.ondot.network.TokenProvider
+import com.dh.ondot.core.network.TokenProvider
 import platform.Foundation.NSBundle
 import platform.Foundation.NSMutableURLRequest
 import platform.Foundation.NSThread
